automate file detail

+++ Core/Service/AssetService.php
@@ -2,10 +2,16 @@
 
 namespace MillenniumFalcon\Core\Service;
 
+use MillenniumFalcon\Core\Nestable\Tree;
 use MillenniumFalcon\Core\Orm\_Model;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Request;
 
 class AssetService
 {
+    const FOLDER_OPEN_MAX_LIMIT = 10;
+
     /**
      * DbService constructor.
      * @param \Doctrine\DBAL\Connection $connection
@@ -13,6 +19,141 @@ class AssetService
     public function __construct(\Doctrine\DBAL\Connection $connection)
     {
         $this->connection = $connection;
+    }
+
+    /**
+     * @return \Pz\Router\InterfaceNode
+     */
+    public function getRoot() {
+        return static::getFolderRoot($this->connection->getWrappedConnection(), 0);
+    }
+
+    /**
+     * @param $currentFolderId
+     * @return \Pz\Router\InterfaceNode
+     */
+    static public function getFolderRoot($pdo, $currentFolderId)
+    {
+        $childrenCount = array();
+        $fullClass = ModelService::fullClass($pdo, 'Asset');
+        $data = $fullClass::data($pdo, array('whereSql' => 'm.isFolder = 1'));
+        foreach ($data as $itm) {
+            if (!isset($childrenCount[$itm->getParentId()])) {
+                $childrenCount[$itm->getParentId()] = 0;
+            }
+            $childrenCount[$itm->getParentId()]++;
+        }
+
+        foreach ($data as &$itm) {
+            if ($itm->getId() == $currentFolderId) {
+                $itm->setStateValue('opened', true);
+                $itm->setStateValue('selected', true);
+            } else {
+                if (isset($childrenCount[$itm->getId()]) && $childrenCount[$itm->getId()] <= static::FOLDER_OPEN_MAX_LIMIT) {
+                    $itm->setStateValue('opened', true);
+                } else {
+                    $itm->setStateValue('opened', false);
+                }
+            }
+        }
+
+        $assetRoot = static::getAssetRoot($pdo, $currentFolderId);
+        $tree = new Tree($data);
+        return $tree->getRootFromNode($assetRoot);
+    }
+
+    /**
+     * @param $pdo
+     * @param $currentFolderId
+     * @return mixed
+     * @throws \Exception
+     */
+    static public function getAssetRoot($pdo, $currentFolderId) {
+        $fullClass = ModelService::fullClass($pdo, 'Asset');
+        $assetRoot = new $fullClass($pdo);
+        $assetRoot->setTitle('Home');
+        $assetRoot->setId('0');
+        $assetRoot->setParentId(null);
+        $assetRoot->setState(array('opened' => true, 'selected' => $currentFolderId == "0" ? true : false));
+        return $assetRoot;
+    }
+
+    /**
+     * @param \PDO $pdo
+     * @param UploadedFile $file
+     * @return JsonResponse
+     * @throws \Exception
+     */
+    static public function processUploadedFile(\PDO $pdo, UploadedFile $file)
+    {
+        $request = Request::createFromGlobals();
+        $parentId = $request->request->get('parentId') ?: 0;
+
+        $originalName = $file->getClientOriginalName();
+
+        $fullClass = ModelService::fullClass($pdo, 'Asset');
+        $rank = $fullClass::data($pdo, array(
+            'select' => 'MIN(m.rank) AS min',
+            'orm' => 0,
+            'whereSql' => 'm.parentId = ?',
+            'params' => array($request->get('parentId')),
+            'oneOrNull' => 1,
+        ));
+        $min = $rank['min'] - 1;
+
+        $orm = new $fullClass($pdo);
+        $orm->setTitle($originalName);
+        $orm->setIsFolder(0);
+        $orm->setParentId($parentId);
+        $orm->setRank($min);
+        $orm->save();
+
+        return static::processUploadedFileWithAsset($pdo, $file, $orm);
+    }
+
+    /**
+     * @param \PDO $pdo
+     * @param UploadedFile $file
+     * @param $orm
+     * @return JsonResponse
+     */
+    static public function processUploadedFileWithAsset(\PDO $pdo, UploadedFile $file, $orm)
+    {
+        static::removeFile($orm);
+        static::removeCaches($pdo, $orm);
+
+        $originalName = $file->getClientOriginalName();
+        $ext = pathinfo($originalName, PATHINFO_EXTENSION);
+
+        $orm->setFileName($originalName);
+        $orm->setFileType($file->getMimeType());
+        $orm->setFileSize($file->getSize());
+        $orm->setFileExtension($file->getClientOriginalExtension());
+
+        $file->move(AssetService::getUploadPath());
+        if (file_exists(AssetService::getUploadPath() . $file->getFilename())) {
+            rename(AssetService::getUploadPath() . $file->getFilename(), AssetService::getUploadPath() . $orm->getId() . '.' . $ext);
+        }
+
+        $info = getimagesize(AssetService::getUploadPath() . $orm->getId() . '.' . $ext);
+        if ($info === false) {
+            $orm->setIsImage(0);
+            $orm->setWidth(null);
+            $orm->setHeight(null);
+        } else {
+            list($x, $y) = $info;
+            $orm->setIsImage(1);
+            $orm->setWidth($x);
+            $orm->setHeight($y);
+        }
+
+        $orm->setFileLocation($orm->getId() . '.' . $ext);
+        $orm->save();
+
+        return new JsonResponse(array(
+            'status' => 1,
+            'orm' => $orm,
+        ));
     }
 
     /**
@@ -36,7 +177,7 @@ class AssetService
      */
     static public function removeFile($asset) {
         $link = static::getUploadPath() . $asset->getFileLocation();
-        if (file_exists($link)) {
+        if (file_exists($link) && is_file($link)) {
             unlink($link);
         }
     }
